requirements.txt
Flask
flask-cors
torch 
torchvision 
pillow 
requests 
tqdm 
tensorboard
gunicorn